requirements.txt
requests==2.28.1
jupyterlab==3.4.7
pyspark==3.2.1
PyYAML==6.0
ipykernel==6.15.0
boto3==1.24.82
pandas==1.3.5
pymongo[srv]
apache-airflow==2.4.3
contractions==0.1.73
flair==0.11.3
langdetect==1.0.9
deep-translator==1.10.1
spacy==3.5.0
nltk==3.8.1
pyarrow==11.0.0
glove==1.0.2
tensorflow==2.11.0
-e .